Разделил классы на два файла Исправил ошибки Исправил тесты
